+++ tests/testthat/test-fix_download_url.R
@@ -6,7 +6,12 @@ download_url <- "public:my_file.pdf"
 
 test_that("URL is correctly fixed", {
   expect_equal(
-    fix_download_url(download_url),
-    "https://datacatalog.worldbank.org//sites//default//filesmy_file.pdf"
+    fix_download_url("public://my_file.pdf"),
+    "https://datacatalog.worldbank.org//sites//default//files//my_file.pdf"
+  )
+
+  expect_equal(
+    fix_download_url("https://datacatalog.worldbank.org//sites//default//files//my file with spaces.pdf"),
+    "https://datacatalog.worldbank.org//sites//default//files//my%20file%20with%20spaces.pdf"
   )
 })